Move count_of_games_by_season and average_goals_per_season methods and tests from GameCollection into StatTracker
