+++ lib/stat_tracker.rb
@@ -16,4 +16,19 @@ class StatTracker
   def games
     GameCollection.new(@game_path)
   end
+
+  def count_of_games_by_season
+    games_per_season = Hash.new{0}
+    games.all.each do |game|
+      games_per_season[game.season] += 1
+    end
+    games_per_season.sort.to_h
+  end
+
+  def average_goals_per_game
+    total_goals = games.all.sum do |game|
+      game.total_score
+    end
+    (total_goals.to_f/games.total_games).round(2)
+  end
 end
