+++ test/team_test.rb
@@ -18,6 +18,7 @@ end
   def test_it_exists
     assert_instance_of Team, @team
   end
+
   def test_initialize
     assert_equal "1", @team.team_id
     assert_equal "23", @team.franchiseId
@@ -26,6 +27,4 @@ end
     assert_equal "Mercedes-Benz Stadium", @team.stadium
     assert_equal "/api/v1/teams/1", @team.link
   end
-
-
 end
