+++ test/test_helper.rb
@@ -7,3 +7,4 @@ require 'pry'
 require_relative '../lib/game'
 require_relative '../lib/stat_tracker'
 require_relative '../lib/team'
+require_relative '../lib/game_collection'
